Working zeroconf mdns implementation
