+++ nmoscommon/mdns/zeroconfEngine/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+from .mdnsEngine import MDNSEngine  # noqa: F401
