+++ nmoscommon/mdns/zeroconfEngine/mdnsSubscription.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+from zeroconf import ServiceBrowser
+
+
+class MDNSSubscription(object):
+
+    def __init__(self, zeroconf, listener, regtype):
+        self.listener = listener
+        self.regtype = regtype + ".local."
+        self.zeroconf = zeroconf
+        self.browser = ServiceBrowser(self.zeroconf, self.regtype, self.listener)
+
+    def close(self):
+        self.browser.cancel()
